rename geolocation_utils.py to geolocation.py
