+++ src/nisarqa/__init__.py
@@ -106,7 +106,7 @@ from .stats_h5_globals import *
 
 # Next import parameters, products, utils, etc.
 from .summary_csv import *
-from .geolocation_utils import *
+from .geolocation import *
 from .raster_classes import *
 from .lonlat import *
 from .parameters import *
